Make identifier decorators and generators more Groovy.
In effect, avoid marking variables as final from abstract classes or
interfaces due to Groovy not allowing us to assign any values to them in
the constructor of concrete implementations.

Also added a missing import to the ModelIdentifier class.

+++ src/groovy/net/biomodels/jummp/core/model/identifier/decorator/FixedDigitAppendingDecorator.groovy
@@ -42,7 +42,7 @@ class FixedDigitAppendingDecorator extends AbstractAppendingDecorator {
     /**
      * Don't pass a suffix below 1 to avoid an IllegalArgumentException.
      */
-    public FixedDigitAppendingDecorator(short order, int suffix, int width)
+    public FixedDigitAppendingDecorator(Integer order, int suffix, int width)
                 throws IllegalArgumentException {
         boolean orderOk = validateOrderValue(order)
         if (!orderOk) {
